Build system update
The Ice package now uses the PACKAGE variable instead of TARGET.
PACKAGE is a fully qualified name, like “cortex::web”. This variable is
used to install the Ice binary in the $CORTEX_HOME/packages/Ice/bin
folder. Additionally, includes will be copied to the
$CORTEX_HOME/packags/Ice/include folder.

The monitor rakefile now includes a USE variable. USE points to an
array of packages that the project depends on. The values in the USE
variable will contain fully qualified names (like PACKAGE). The build
system will use these to link with the correct library and provide the
correct include path.